+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [[tbd]] - 2024-10-17
+
+### Fixed
+- LT-5719: Avoid using custom basic properties when publishing messages
+
 ## 8.6.2 - 2024-10-17
 
 ### Fixed